Add files via upload
This is the only required file however, you must have python3 installed and BeautifulSoup4. Once you have both installed, change your directory to where the file is saved. Run with : python3 webscrape.py . The script will run through each page of the specials then output all info to a .csv file in the same directory which can be viewed in Exel or Numbers.